refactor(messaging): split platform_adapter.go by responsibility (SRP)
Split the 219-line platform_adapter.go into 4 single-responsibility
files: platform_types.go (PlatformType + constants + ExtractPlatformKeys),
platform_interfaces.go (HubInterface, HandlerInterface, SessionStarter,
PlatformAdapterInterface, CronResultSender), platform_registry.go
(AdapterBuilder, Register, New, RegisteredTypes), and platform_adapter.go
(PlatformAdapter struct + methods only).

Refs #337
